vehicle details added and input validation
